Give the command modifiers one home in types/ex_cmds.rs
`CMOD_*` was 14 names in 139 declarations over 27 files, spread across
*eleven* different `C2Rust_Unnamed_NN` aliases — each of which means
something unrelated in every other file that has one. The rows now sit
beside `cmdmod_T`, under a name that says what they are: `CmdModFlags`.

They stay `c_uint`. `cmod_flags` is a `c_int` and the honest type for the
flags is `c_int` too, but every transpiled test site spells
`CMOD_X as ::core::ffi::c_int`, so retyping now would turn ~200 casts
into same-type casts and grow the clippy count. That retype belongs with
the rewrite that deletes the casts.

Six use sites moved: search/command.rs held one of B11's
`const X: T = super::X as T;` alias rows (`consolidate.py` refuses
those), and five sites that had been reading a locally-`c_int`-typed
copy now cast at the point of use, the way their siblings already do.
Nothing else changed but `use` lines.

175 declarations dropped, -112 capped lines; `unsafe `, `no_mangle`,
`static_mut` and clippy unchanged.

+++ crates/nvim/src/nvim/mark/adjust.rs
@@ -11,6 +11,7 @@ use core::ffi::{c_int, c_uint};
 use core::ptr;
 
 use super::*;
+use crate::src::nvim::types::CMOD_LOCKMARKS;
 
 /// `mark.c`'s `ONE_ADJUST` family: the line-number rewrite every mark store
 /// gets when lines are inserted, deleted or moved. Marks in `line1..=line2`
@@ -377,7 +378,9 @@ pub unsafe extern "C" fn mark_col_adjust(
         col_amount,
         spaces_removed,
     };
-    if col_amount == 0 && lnum_amount == 0 || (*cmdmod.ptr()).cmod_flags & CMOD_LOCKMARKS != 0 {
+    if col_amount == 0 && lnum_amount == 0
+        || (*cmdmod.ptr()).cmod_flags & CMOD_LOCKMARKS as c_int != 0
+    {
         return;
     }
     let mut i: c_int = 0;
